refactor + lint + build fix
