+++ .gitignore
@@ -45,3 +45,4 @@ yarn-error.log*
 /.idea/
 
 /lap-student-tracker-staging-firebase-adminsdk-b957b-058bd73482.json
+/.env.staging
