+++ components/SingleClassView/Modules/Module.tsx
